# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Client API",
        "description": "API documentation for QuickDials application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/business/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard data",
                "description": "Fetch user/business dashboard overview. Requires Bearer token.",
                "operationId": "bc2333acb511aa0be013f2840ec8c796",
                "responses": {
                    "200": {
                        "description": "Dashboard data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_ads": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "active_ads": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "pending_ads": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "messages": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "profile_completed": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "last_login": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-09-04T10:20:30Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/business/get-discussion": {
            "get": {
                "tags": [
                    "Discussion"
                ],
                "summary": "Get all discussions",
                "description": "Fetch all discussions. Requires Bearer token.",
                "operationId": "4d0c9fe4a12c6c4be0b7dcb9e2e3ea2e",
                "responses": {
                    "200": {
                        "description": "List of discussions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "How to improve SEO for my business?"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "Amit Sharma"
                                                    },
                                                    "author_id": {
                                                        "type": "integer",
                                                        "example": 42
                                                    },
                                                    "replies_count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "last_reply_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-09-04T14:30:00Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-09-03T09:15:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "BearerAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    }
}